Remove readings older than 24 hours
